requirements.txt
streamlit>=1.33.0
openai>=1.30.1
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.8.0  #For vector similarity